mem_opt

+++ pageReplace.cpp
@@ -5,7 +5,7 @@
 #define GET_PAGE(x) ((x)>>MAX_PAGE)
 
 bool lastuse[MAX_PHY_PAGE];
-long map[NUM_OF_PAGE];
+int map[NUM_OF_PAGE];
 
 //second chance:
 void pageReplace(long* physic_memory, long nwAdd)
